Updated openstack/openstack Project: openstack/manila  544cb2aabe79c64842f1f629f13a320c0e2ae480
Fix rpcapi identifiers for better readability

In manila/share/rpcapi and manila/scheduler/rpcapi
files, 'cctxt' is used to receive the prepared rpcapi
client call context from oslo_messaging. The purpose
of this identifier is not immediately discernible
from the other 'ctxt' used in the files (that denotes
request context).

Fix it to enhance readability.

TrivialFix

Change-Id: I54c614d9e963da23a0aeb92cb196fddf5ec751e6