Updated openstack/openstack Project: openstack/neutron  5b7c6d9e89bb35108e6f74aaf90a376bafc0a6cd
Pecan: Always associate plugins with resource

with this patch, the logic for associating a resource with a plugin
is now executed for every resoruce. This will avoid requiring the
method get_pecan_controllers in extensions descriptors to deal with
this. This item of work is required for a speedy "Pecanization" of
existing extensions, in particular the 'router' extension.

The routine for finding a plugin for a resource has been modified to
allow special treatment of the 'quotas' extension. This extension
indeed is declared as supported by plugins (usually the core one),
but the plugin does not implement relevant methods as quota management
is performed by a distinct driver.

Further, NeutronPecanController's plugin attribute has been
converted into a property which loads the value from NeutronManager
if not yet defined. Indeed in some cases the plugin might be
instantiated after the controller instance is created.

Closes-Bug: #1528510

Change-Id: Ibbfec8fd53855641bd21dec8ef824d5741dfebea