Updated openstack/openstack Project: openstack-infra/project-config  8613c046e980e38e87e82b78d4a042e7c5938749
run trove experimental run scenario tests

Trove experimental jobs are running old tests that have been changed.
This change is to run the correct group of scenario tests for redis
and mysql. The tests are concidered experimental still because they
are still a work in progress to replace the existing int-tests that
trove uses for testing mysql. This change also helps us setup the
scenario tests so that we can verify other datastores supported in
trove in the future.

Change-Id: If6582ad2497e85b66ff5c90d8317004f059fbea7