Updated openstack/openstack Project: openstack/keystone  695ea9a857f3dff5eff2c4d48c0e7edbb00821be
Modify rules for domain specific role assignments

Building on the support for domain specific roles, this final
patch in the series adjusts the policy rules for grants using
domain specific roles in the v3cloudsample.

In addition, a release note is provide for the domain specifc
role functionality.

Implements: blueprint domain-specific-roles
Change-Id: I941611237ddd48cb87337bcebd61cfda52479f87