Updated openstack/openstack Project: openstack-dev/devstack  886410eadeaf5366e51bd31677a3b49b34eb391c
Add guide on running devstack in lxc container

Running OpenStack in a container can be a useful workflow for developers.
The primary benefits are faster performance and lower memory overhead
while still providing a suitable level of isolation.

The guide walks the user through procedure for configuring an LXC container
and deploying OpenStack in it using devstack. It also discusses the limitations
of this setup - particularly related to cinder.

Change-Id: I2e0921fd118cfe98cef86ba110a94b3edccf9a29