Updated openstack/openstack Project: openstack/zaqar  b24be72fde836d543b6a2e45ba08f389d2fcb079
Don't return links if subscriptions are empty

The commit I08dd0f675bcd25cd63e17a0aa9779895bd401e32 try to fix
the links format when a queue's subscriptions are empty. But it
doesn't work. Because the "results" which storage returns will
never be None.

This patch fix the nit and also add a test to ensure it works.

Change-Id: Ib7fb0b6bfccc5e913ff618257f07903f692201f1