Updated openstack/openstack Project: openstack/keystone  8a8964b59bbca002ed175d764190aefc89570321
Allow project domain_id to be nullable at the manager level

Since the domain_id column is part of a uniqueness constraint and
it might now contain a nullable value, we actually store a special
value (as opposed to None) in the sql attribute so as to still use
the sql uniqueness constraint to ensure we do not have race
conditions in multi-process keystone configurations. So as to
continue to support a FK on the project.domain_id column, we create
a hidden domain ("the root of all domains"). In preparation for the
move to projects acting as domains, we also create a matching project
for this root. These special rows, and indeed the special domain_id
value stored, are entirely hidden within the sql driver and are not
exposed to the manager layer and above. The ability to store a project
domain_id of None is also added to the wrapper for legacy drivers.

In addition, one incorrectly written test relating to upcoming support
of projects as a domain had to be modified, since it started
passing prematurely, and the (now unused) driver base class method
for setting a domain_id is moved to the legacy V8 class.

Follow-on patches will build on this enable projects acting as domains,
which will have a null domain_id.

Co-Authored-By: Henry Nash <henryn@linux.vnet.ibm.com>
Partially-Implements: bp reseller

Change-Id: I4e5b9f3978482082c23f729d29bbdc701bc7c473