Updated openstack/openstack Project: openstack/keystone  8d7d49edaf99f0915ad7226b02b3d51218833cd4
Verify project unique constraints for projects acting as domains

With the creation of projects acting as domains, represented
by the flag 'is_domain', the fields 'name' and 'domain_id'
should still be enough to uniquely identify a project, since
all projects acting as a domain have a null domain_id.

This patch adds testing to validate this, as well as provides
clearer error messages for when there is a clashing name.

Co-Authored-By: Rodrigo Duarte <rodrigods@lsd.ufcg.edu.br>
Co-Authored-By: Henrique Truta <henrique@lsd.ufcg.edu.br>
Co-Authored-By: Henry Nash <henryn@linux.vnet.ibm.com>

Partially-Implements: bp reseller

Change-Id: I5040fc780962382d18d18fecd9b51ed1d2759a2b