Updated openstack/openstack Project: openstack/ironic  a70b5365d34746d0e5f5d6f932829625d73c17cf
During cleaning, store clean step index

During cleaning, we keep the list of clean steps in the node's
driver_internal_info['clean_steps']. node.clean_step is the current
clean step being performed and is a copy of a clean step in the list.
This change adds a driver_internal_info['clean_step_index'] that
is an index into the list; it corresponds to node.clean_step.

This addresses an issue in the code where a clean step that occurs
more than once in the list would cause cleaning to loop forever.

Change-Id: Ibbb3a0239bb288eb3d2d3c60fc8ddd1a16e2927a
Closes-Bug: #1536276