Updated openstack/openstack Project: openstack/keystone  922833c362aba22fd7614414ea68fbbc78cb020d
Add is_domain filter to v3 list_projects

In preparation for supporting projects acting as a domain, we
add filtering to the list_projects API, with a default of
is_domain=False. This ensures that the collection returned
from an unfiltered call at the API level will be unchanged
as compared to earlier releases.

Co-Authored-By: Henrique Truta <henrique@lsd.ufcg.edu.br>
Co-Authored-By: Henry Nash <henryn@linux.vnet.ibm.com>

Partially-Implements: bp reseller
Change-Id: Ia53f73e578834c191d04ede2b76a104c29b14678