Updated openstack/openstack Project: openstack-infra/project-config  c6d1648dd1d6bc64e7c63aa4348691ee71d02bcf
Use pip install of dib for testing

Rather than running from the checkout of the source tree, we should
also be testing the install/distribution part of the package by pip
installing.

This sets up some experimental jobs to make sure this works, with the
idea of converting the main jobs when it does.

Change-Id: I161f58d19e57cdc985d49d8490f57cae2a441732