Updated openstack/openstack Project: openstack-infra/project-config  9645e1d705df199f057568e628cbbf3d38405ed4
Fix tag publishing of server projects

Projects using openstack-server-publish-jobs publish documents both from
post and release pipelines. Now a tag on stable branch will publish the
documents both to developer/$project/$tag and developer/$project. This
overrides the documents published in the post pipeline - replacing
master documents with stable documents.

The publishing to both locations is correct for projects that only
publish as part of releases but not for those useing
openstack-server-publish-jobs.

To fix this:
Create a new job '{name}-docs-tags-only' that adds a new parameter to
run-docs script. If that parameter is passed, tag publishing will only
happen to developer/$project/$tag . Add the new job to the
openstack-publish-jobs so that all projects can use it.

Adapt openstack-server-publish-jobs in zuul to use the new job.

This change also adds constraint version of the publishing job. A new
builder docs-tags-only-env has beend added for the new jobs.

Change-Id: If0d000358f17a7d9ea0b8c3e3aada1c16291b3ee