Updated openstack/openstack Project: openstack/heat  834ac4a04171a3b1e619962ebfbd02034d36ec22
Ignore connection error when testing load balancer

In the integration test with Neutron Load Balancer, we sometimes fail to
connect to the Load Balancer when testing that the stack deployed
successfully. Let's ignore those errors so that we retry for a bit in
case it takes some time to be deployed.

Change-Id: Ic97bd83fe56dc52fb49848d93e40cb57d1550a1f
Closes-Bug: #1545587