Updated openstack/openstack Project: openstack/nova  ee12a3a82efc9e78d10f09154563df2abbbf15ad
libvirt: make live_migration_uri flag dependent on virt_type

The default value for the "live_migration_uri" flag should be
dependent on the "virt_type" flag, as the "connection_uri" flag is. This
way, an operator can set the "virt_type" flag without the need to check
for each individual uri.

DocImpact: Changed the default value of the "live_migration_uri" flag,
that now is dependent on the "virt_type".
Closes-Bug: #1298242
Change-Id: Id54f7bdfa14a19da41da554b13ba9496ee525c71