Updated openstack/openstack Project: openstack/python-novaclient  9774405f8e6edb5aad83866c7c94226ab15ab4cc
Allow to specify a network for functional tests

Currently the tests just attach instances to the first network returned
by "net-list". That might however not be the right thing in some
environments. This change allows to override the default network via the
environment variable "OS_NOVACLIENT_NETWORK". If not specified the
test will fallback to the old behaviour and just use the first network.

Closes-Bug: #1510975
Change-Id: Ie682111127584a33d8e96377d812d3a6352c760d