Updated openstack/openstack Project: openstack/python-keystoneclient  9097481e81f5b45310f4f5228369325e3bbc0f28
Implied Roles

While the entity for an inference rule should be thought of as a
resource, the rules are essentially relationships between roles.
The `implied_role` API is linked with the role API, and thus the
client functions are part of v3/role.py.  However, it does not
map completely cleanly to the Crud baseclass, and requires
some custom URL generation.

Change-Id: I80a40e88b571fe9b0eca3af8b705ea79f28eb904