Updated openstack/openstack Project: openstack/neutron  171351d5fb55f96e9ed9e56eab755f2717ee67b0
Adopt Grafana to plot Neutron Failure Rates

Maintaining these dashboards in tree is a bit painful, the queries
can be brittle, the URLs are not human readable, and furthermore the
graphs are not interactive (e.g. zoomable).

We should adopt Grafana, and to see how this looks like, please check
tempest failure rate inline.

Change-Id: I422fc04015fa50e2176e41128adc41d7a22fe02b