Updated openstack/openstack Project: openstack-infra/project-config  8f14307e1dad3be14b32e27488f020fcd29d3aab
Common template for Centos & Fedora devstack jobs

Consolidate these jobs in a single "platform" template that is pretty
generic but allows us to specify the node to run on (might be useful
to other nodes in near future, such as next Ubuntu LTS and Debian).

I wasn't sure about "platform" -- I also considered something like
"altnode" or just "node" to differenitate it from the main jobs.
"Platform" seems to make it clear it's a different environment

To maintain the status-quo of running serial smoke-tests on centos7
I have added an argument for that (working on getting rid of that
requirement).

We've got "-nv" suffix now, so remove the non-voting regexes.

This also restores Fedora 23 to non-voting gate check for devstack --
all our blocking issues with that are worked out and the nodes are
available in multiple providers.

Change-Id: I3a58144426ff4e904c316210b647379b1ba7305d