Updated openstack/openstack Project: openstack-dev/devstack  80ea24cbbb98fd6bdc5dff2fa4bbfc074a1b38b5
Always strip package comments

In the "I can't believe I missed this" category -- the existing strip
method removes shortest match (%); which fails when you put another #
in the comment (like "refer to bug#1234").  Change to the longest
match which should strip everything from the first "#" to the end
(since that's going to be the longest).

Change-Id: I47f5e710ebd87b0f54549732e7d64cf42c7a6b65