Updated openstack/openstack Project: openstack-infra/storyboard-webclient  17a76768f5e6b8d1c037d00b168944fe0dee816c
Added bower install to post-install hook

This patch adds an invocation to `bower install` to the project's
postinstall hook. The end result is that developers no longer
have to explicitly invoke both `npm install` and `bower install`.
It also brings this project closer in line with the OpenStack npm
job templates.

Change-Id: Ib5fb72c630a97506f3b3bceca6c98302671ccf28