Updated openstack/openstack Project: openstack/keystone  71fe1886e1834d9fb05d8d409e6606a79b310ef5
Allow user list without specifying domain

With a single domain environment, users can be listed without
specifying a domain. When moving to a multiple domain environment,
this remains true for domain-scoped tokens but not for project-scoped
tokens. Project-scoped tokens currently only work if the domain_id
query parameter is specified. This has been a source of pain to many
users, and is unnecessary. Just as the desired domain is assumed to be
that to which the token is scoped when the token is domain-scoped,
keystone can assume the desired domain is that of the project's domain
when the token is project-scoped.

Change-Id: I1d06935c06661109a523c5b4547ff01f23235a89
Closes-Bug: 1479578