Updated openstack/openstack Project: openstack/horizon  9d62f141fec98dd653804b6c2f6ef84e2f77605b
Hide delete volume if it has snapshot

If the volume has a snapshot, it is not allowed to
delete it. In tables the delete action is hidden but
if we go to the volume detail page, the delete action
is available.

This patch hides the delete volume on detail page too.

Change-Id: I4d8690b035dedd7ebcacb3479d346cfb3fb324f1
Closes-Bug: #1546423