Updated openstack/openstack Project: openstack/nova  43e39c5a2ed407f0f6587ae53e58d66ceea4dc83
Remove support for integer ids in compute_api.get

Because of the in tree ec2 api we used to need to easily handle both
integer and uuid ids as if they were the same thing. This led to
making compute_api.get take either and try to do the right thing. This
unfortunately meant we exposed the indexed integer ids over the
OpenStack API, which was never intended.

... and turns out to totally explode because of the way we cache
requests for extensions to post process them. The hide_address
extension shows a nice bit of exploding which has been noticed by our
users.

Remove the int id facility. Remove the test for it. Update the other
tests that were sneaking in and using id==1 like very naughty folks
(there will be coal in your stocking this year).

Closes-Bug: #1522536

Change-Id: Ibff4cebe71714ac9c470337c2ee5f57df4343829