Updated openstack/openstack Project: openstack/nova  69e64aa8630067fa7701f95bbeaed6c26571f1d5
Add better help text to scheduler options (6)

The previous commit moved all the scheduler options to a central
location. This commit adds additional help text to those options.
Because of the large number of options, the improved help text will be
split into several patches; this is patch 6 of 7.

Blueprint centralize-config-options

Change-Id: Iec25eb2072d41f2f04f9aa6ac2cc6b4e7aea4023