Updated openstack/openstack Project: openstack/cinder  6ba8d81e8be09be0f10c56e7a1788edff168f164
Remove useless get_replication_updates driver call

We had the call in the manager as part of a periodic,
it did nothing at all.  Some drivers had comments "no idea
what to do with this", even better we had unit tests that
faked out a call that returned nothing and asserted it was
nothing. ?!?

Let's just remove it

Change-Id: I3c1ef6065e73d90912d7f765d581add9bc704c7b
Closes-Bug: #1547666