Updated openstack/openstack Project: openstack/cinder  ee73d93c8b66d02a7ef33c36443143c40b1cba6f
Wait for periodic tasks to stop on exit

When a Service in Cinder exits it stops all operations and waits for
them to complete, and it also stops all periodica tasks created with
FixedIntervalLoop, but it doesn't wait until these are completed.

This patch will wait for all periodic tasks that received the stop
request without raising an exception before exiting.

Change-Id: I984b8aebd40f9482f0cf8098146d1ccd635dc1d6
Closes-Bug: #1546234