Updated openstack/openstack Project: openstack/oslo.messaging  bf6c31a5ed02b4463b27d7b9f62cde7cdc294141
Option rpc_response_timeout should not be used in zmq driver

Client option rpc_response_timeout means "Seconds to wait for a response
from a call"[1], but it is used by cast in zmq driver. And it's not the
right way that underlying zmq driver uses an upper client option.

So rpc_response_timeout should not be used in zmq driver, and let's just
remove it since rpc client has passed the parameter[2].

ref:
[1]: https://github.com/openstack/oslo.messaging/blob/master/oslo_messaging/rpc/client.py#L37
[2]: https://github.com/openstack/oslo.messaging/blob/master/oslo_messaging/rpc/client.py#L150

Change-Id: I86fe4d1f18faf8d4af666301b00e5907dce4417b